# backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/google:
    get:
      operationId: googleAuthStart
      tags: [google]
      summary: Start Google OAuth consent flow
      description: Returns the Google consent URL that a lawyer must visit to connect their Google Calendar.
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Authorization URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleConnectResponse"
        "503":
          description: Google Calendar not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/google/status:
    get:
      operationId: googleAuthStatus
      tags: [google]
      summary: Check Google Calendar connection status
      description: Returns whether the given user has connected their Google Calendar.
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleStatusResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    GoogleConnectResponse:
      type: object
      properties:
        authorizationUrl:
          type: string
      required:
        - authorizationUrl
    GoogleStatusResponse:
      type: object
      properties:
        connected:
          type: boolean
      required:
        - connected
...